add support for audit log enabled and compliance enabled
