+++ legacy/securityconfig_v6/audit.yml
@@ -1,4 +1,7 @@
 config:
+  # enable/disable auditlog
+  enabled: true
+
   audit:
     # rest
     enable_rest: true
@@ -24,6 +27,9 @@ config:
     exclude_sensitive_headers: true
 
   compliance:
+    # enable/disable compliance
+    enabled: true
+
     # configs
     internal_config: true
     external_config: true
