+++ src/main/java/com/amazon/opendistroforelasticsearch/security/auditlog/impl/AuditLogImpl.java
@@ -21,6 +21,8 @@ import java.security.AccessController;
 import java.security.PrivilegedAction;
 import java.util.Map;
 
+import com.amazon.opendistroforelasticsearch.security.auditlog.config.AuditConfig;
+import com.amazon.opendistroforelasticsearch.security.compliance.ComplianceConfig;
 import org.elasticsearch.SpecialPermission;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
@@ -39,11 +41,15 @@ import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportRequest;
 
 import com.amazon.opendistroforelasticsearch.security.auditlog.routing.AuditMessageRouter;
+import org.greenrobot.eventbus.Subscribe;
 
 public final class AuditLogImpl extends AbstractAuditLog {
 
 	private final AuditMessageRouter messageRouter;
-	private final boolean enabled;
+	private final Settings settings;
+	private final boolean dlsFlsAvailable;
+	private final boolean messageRouterEnabled;
+	private volatile boolean enabled;
 
 	public AuditLogImpl(final Settings settings,
 			final Path configPath,
@@ -51,7 +57,7 @@ public final class AuditLogImpl extends AbstractAuditLog {
 			final ThreadPool threadPool,
 			final IndexNameExpressionResolver resolver,
 			final ClusterService clusterService) {
-		this(settings, configPath, clientProvider, threadPool, resolver, clusterService, null);
+		this(settings, configPath, clientProvider, threadPool, resolver, clusterService, null, true);
 	}
 
 	public AuditLogImpl(final Settings settings,
@@ -60,12 +66,18 @@ public final class AuditLogImpl extends AbstractAuditLog {
 						final ThreadPool threadPool,
 						final IndexNameExpressionResolver resolver,
 						final ClusterService clusterService,
-						final Environment environment) {
+						final Environment environment,
+						final boolean dlsFlsAvailable) {
 		super(settings, threadPool, resolver, clusterService, environment);
+		this.settings = settings;
+		this.dlsFlsAvailable = dlsFlsAvailable;
+		if (!dlsFlsAvailable) {
+			log.debug("Changes to Compliance config will ignored because DLS-FLS is not available.");
+		}
 		this.messageRouter = new AuditMessageRouter(settings, clientProvider, threadPool, configPath);
-		enabled = this.messageRouter.isEnabled();
+		this.messageRouterEnabled = this.messageRouter.isEnabled();
 
-		log.info("Message routing enabled: {}", this.enabled);
+		log.info("Message routing enabled: {}", this.messageRouterEnabled);
 
 		final SecurityManager sm = System.getSecurityManager();
 
@@ -95,9 +107,18 @@ public final class AuditLogImpl extends AbstractAuditLog {
 
 	}
 
+	@Subscribe
+	public void onAuditConfigChanged(final AuditConfig auditConfig) {
+		enabled = auditConfig.isEnabled() && messageRouterEnabled;
+		onAuditConfigFilterChanged(auditConfig.getFilter());
+		if (dlsFlsAvailable) {
+			onComplianceConfigChanged(ComplianceConfig.from(auditConfig.getCompliance(), settings));
+		}
+	}
+
 	@Override
-	protected void enableRoutes(Settings settings) {
-		if (messageRouter.isEnabled()) {
+	protected void enableRoutes() {
+		if (messageRouterEnabled) {
 			messageRouter.enableRoutes(settings);
 		}
 	}
