+++ src/main/java/com/amazon/opendistroforelasticsearch/security/dlic/rest/validation/AuditValidator.java
@@ -14,6 +14,7 @@ public class AuditValidator extends AbstractConfigurationValidator {
                           final Object... param) {
         super(request, ref, esSettings, param);
         this.payloadMandatory = true;
+        this.allowedKeys.put(AuditConfig.Key.ENABLED, DataType.BOOLEAN);
         this.allowedKeys.put(AuditConfig.Key.AUDIT, DataType.OBJECT);
         this.allowedKeys.put(AuditConfig.Key.COMPLIANCE, DataType.OBJECT);
     }
