+++ src/test/resources/auditlog/audit.yml
@@ -3,6 +3,9 @@ _meta:
   config_version: 2
 
 config:
+  # enable/disable auditlog
+  enabled: true
+
   audit:
     # rest
     enable_rest: false
@@ -24,6 +27,9 @@ config:
     exclude_sensitive_headers: false
 
   compliance:
+    # enable/disable compliance
+    enabled: true
+
     # configs
     internal_config: true
     external_config: false
