+++ src/test/resources/legacy/securityconfig_v6/audit.yml
@@ -1,4 +1,6 @@
 config:
+  # enable/disable auditlog
+  enabled: true
 
   audit:
     # rest
@@ -21,6 +23,9 @@ config:
     exclude_sensitive_headers: false
 
   compliance:
+    # enable/disable compliance
+    enabled: true
+
     # configs
     internal_config: true
     external_config: false
